fix(): Change the order of cmd-nsc-init execution (#13)
Placed the cmd-nsc-init init container before all other init containers.
This is done to make the nsc-init to run before the istio init container.
The istio init container installs iptable rules to route all the
container traffic to the istio proxy sidecar. But until the istio sidecar
container is in running state, all traffic blackholes. So we need the
cmd-nsc-init container to run first and establish connection to the vl3
nse. This is needed only if the connection between the cmd-nsc-init and
the nsmgr is on a network socket, tcp or udp.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>